Chore: Removed the whitespaces and format the code using IntelliJ default code formatter
