+++ muser/pom.xml
@@ -2,7 +2,7 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>edu.usf.cutr</groupId>
+	<groupId>edu.usf.sas.pal.muser</groupId>
 	<artifactId>muser</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<build>
@@ -50,7 +50,7 @@
 					<transformers>
 						<transformer
 							implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-							<mainClass>edu.usf.cutr.mfe.ProcessorMain</mainClass>
+							<mainClass>edu.usf.sas.pal.muser.ProcessorMain</mainClass>
 						</transformer>
 						<!-- files overwrite each other and geotools does not function without 
 							this. http://docs.geotools.org/latest/userguide/faq.html#how-do-i-create-an-executable-jar-for-my-geotools-app -->
