--- muser/src/main/java/edu/usf/cutr/mfe/exception/FirebaseFileNotInitializedException.java
@@ -1,7 +1,0 @@
-package edu.usf.cutr.mfe.exception;
-
-public class FirebaseFileNotInitializedException extends Exception {
-	public FirebaseFileNotInitializedException() {
-		super("Firebase file is not initialized.");
-	}
-}
