+++ muser/src/main/java/edu/usf/sas/pal/muser/exception/FirebaseFileNotInitializedException.java
@@ -1,0 +1,7 @@
+package edu.usf.sas.pal.muser.exception;
+
+public class FirebaseFileNotInitializedException extends Exception {
+    public FirebaseFileNotInitializedException() {
+        super("Firebase file is not initialized.");
+    }
+}
